ユースケースの分離
